third_party/blink/web_tests/platform/generic/fast/js/function-name-expected.txt

Test for function.name

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS (function f() {}).name is 'f'
PASS delete (function f() {}).name is true
PASS (function() {}).name is ''
PASS Math.name is undefined
PASS Error.name is 'Error'
PASS String.prototype.charAt.name is 'charAt'
PASS document.getElementById.name is 'getElementById'
PASS successfullyParsed is true

TEST COMPLETE
